Rename and start to clean up command
